Add manifest version file
This is a crude hack because I'm not using a build system, but
need to have a running version number to keep the manifest file
seem fresh to the browser. It works, so in other words it's
a fantastic solution. Maybe I'll install gulp when I get a better
wi-fi connection.